+++ module/apps/PinCushion.js
@@ -548,7 +548,7 @@ export class PinCushion {
       </div>
     `);
 
-    // TODO 
+    // TODO
     /*
     <div class="form-group">
       <label